Updated problems 2 - 6.
Changed all int to unsigned or unsigned long.

+++ problem_003.h
@@ -1,6 +1,6 @@
 #ifndef PROBLEM_003_H_
 #define PROBLEM_003_H_
 
-int problem_003();
+unsigned long problem_003();
 
 #endif
